tests(api-core): move power shelf tests to integration and DB suites
