+++ crates/api-core/src/tests/mod.rs
@@ -79,7 +79,6 @@ mod network_segment_find;
 mod network_segment_lifecycle;
 mod nvl_instance;
 mod nvl_logical_partition;
-mod power_shelf;
 mod power_shelf_health;
 mod power_shelf_state_controller;
 mod preingestion_dpu_nic_mode;
