+++ crates/api-db/src/test_support/mod.rs
@@ -15,5 +15,8 @@
  * limitations under the License.
  */
 
+#[cfg(test)]
 pub(crate) mod expected_host;
+#[cfg(test)]
 pub(crate) mod network_segment;
+pub mod power_shelf;
